requirements.txt
torch >= 2.1.0
pytorch-lightning >= 2.2.4
click >= 8.1.7
numpy >= 1.26.4
matplotlib >= 3.9.0

# data version control
dvc >= 3.50.2
dvc-gdrive >= 3.0.1

mlflow >= 2.13.0
hydra-core >= 1.3.2